conversation-memoy-in-langchain/pyproject.toml
[project]
name = "memoryinlangchain"
version = "0.1.0"
description = "Implementing Conversational Memory in LLMs with LangChain"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.26",
    "langchain-groq>=0.3.4",
    "langgraph>=0.4.8",
    "mlflow>=3.1.0",
    "python-dotenv>=1.1.1",
]
